pid width

+++ src/infoLoggerAdminDB.cxx
@@ -208,7 +208,7 @@ int main(int argc, char* argv[])
   }
 
   std::string sqlTableDesriptionMessages =
-    "(severity char(1), level tinyint unsigned, timestamp double(16,6), hostname varchar(32), rolename varchar(32), pid smallint \
+    "(severity char(1), level tinyint unsigned, timestamp double(16,6), hostname varchar(32), rolename varchar(32), pid mediumint \
     unsigned, username varchar(32), `system` varchar(32), facility varchar(32), detector varchar(32), `partition` varchar(32), run int unsigned, errcode int unsigned, \
     errline smallint unsigned, errsource varchar(32), message text, index ix_severity(severity), index ix_level(level), index ix_timestamp(timestamp), index \
     ix_hostname(hostname(14)), index ix_rolename(rolename(20)), index ix_system(`system`(3)), index ix_facility(facility(20)), index ix_detector(detector(8)), index \
@@ -248,7 +248,7 @@ int main(int argc, char* argv[])
   if (optCreate) {
     log.info("Creating infoLogger table");
     // prepare insert query corresponding to 1st protocol definition
-    // create table messages(severity char(1), level tinyint unsigned, timestamp double(16,6), hostname varchar(32), rolename varchar(32), pid smallint unsigned, username varchar(32), system varchar(32), facility varchar(32), detector varchar(32), `partition` varchar(32), run int unsigned, errcode int unsigned, errline smallint unsigned, errsource varchar(32), message text, index ix_severity(severity), index ix_level(level), index ix_timestamp(timestamp), index ix_hostname(hostname(14)), index ix_rolename(rolename(20)), index ix_system(system(3)), index ix_facility(facility(20)), index ix_detector(detector(8)), index ix_partition(`partition`(10)), index ix_run(run), index ix_errcode(errcode), index ix_errline(errline), index ix_errsource(errsource(20)));
+    // create table messages(severity char(1), level tinyint unsigned, timestamp double(16,6), hostname varchar(32), rolename varchar(32), pid mediumint unsigned, username varchar(32), system varchar(32), facility varchar(32), detector varchar(32), `partition` varchar(32), run int unsigned, errcode int unsigned, errline smallint unsigned, errsource varchar(32), message text, index ix_severity(severity), index ix_level(level), index ix_timestamp(timestamp), index ix_hostname(hostname(14)), index ix_rolename(rolename(20)), index ix_system(system(3)), index ix_facility(facility(20)), index ix_detector(detector(8)), index ix_partition(`partition`(10)), index ix_run(run), index ix_errcode(errcode), index ix_errline(errline), index ix_errsource(errsource(20)));
     std::string sqlQuery;
     sqlQuery = "create table " INFOLOGGER_TABLE_MESSAGES + sqlTableDesriptionMessages;
     if (mysql_query(&db, sqlQuery.c_str())) {
